+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - Now relies on pwix:modal v 1.6.0 (todo #1)
     - Clicking on the backdrop no more close the dialogs (todo #3)
     - Update to pwix-modal 1.7.0 et set version requirement
+    - Bump Meteor version requirement
 
 ### 1.5.0